Split tests across several components
