+++ tox.ini
@@ -31,6 +31,8 @@ set_env =
     dynamic-allocation: TESTFILE=test_dynamic_allocation.py
     observability: TESTFILE=test_observability.py
     tls: TESTFILE=test_tls.py
+    auth: TESTFILE=test_auth.py
+    relation: TESTFILE=test_relation.py
 pass_env =
     PYTHONPATH
     CHARM_BUILD_DIR
@@ -82,7 +84,7 @@ commands =
 
 
 
-[testenv:integration-{charm,trust,ha,upgrade,external-access,tls,dynamic-allocation,observability,iceberg,metastore}]
+[testenv:integration-{charm,trust,ha,upgrade,external-access,tls,dynamic-allocation,observability,iceberg,metastore,auth,relation}]
 description = Run integration tests
 set_env =
     {[testenv]set_env}
